+++ PennMobileShared/Extensions/Collection+Extensions.swift
@@ -1,0 +1,22 @@
+//
+//  Collection+Extensions.swift
+//  PennMobileShared
+//
+//  Created by Josh Doman on 12/13/16.
+//  Copyright © 2016 Josh Doman. All rights reserved.
+//
+
+import Foundation
+
+public extension Collection {
+    /// Return a copy of `self` with its elements shuffled
+    func shuffle() -> [Iterator.Element] {
+        var list = Array(self)
+        list.shuffleInPlace()
+        return list
+    }
+    
+    var random: Iterator.Element? {
+        return shuffle().first
+    }
+}
