+++ PennMobileShared/Extensions/Locale+Extensions.swift
@@ -1,0 +1,13 @@
+//
+//  Locale+Extensions.swift
+//  PennMobileShared
+//
+//  Created by Josh Doman on 12/13/16.
+//  Copyright © 2016 Josh Doman. All rights reserved.
+//
+
+import Foundation
+
+public extension Locale {
+    static let enUS = Locale(identifier: "en_US_POSIX")
+}
