+++ PennMobileShared/Extensions/Logger+Extensions.swift
@@ -1,0 +1,16 @@
+//
+//  Logger+Extensions.swift
+//  PennMobileShared
+//
+//  Created by Josh Doman on 12/13/16.
+//  Copyright © 2016 Josh Doman. All rights reserved.
+//
+
+import Foundation
+import OSLog
+
+public extension Logger {
+    init(category: String) {
+        self.init(subsystem: Bundle.main.bundleIdentifier ?? "Penn Mobile", category: category)
+    }
+}
