+++ PennMobileShared/Extensions/NSMutableAttributedString+Extensions.swift
@@ -1,0 +1,45 @@
+//
+//  NSMutableAttributedString+Extensions.swift
+//  PennMobileShared
+//
+//  Created by Josh Doman on 12/13/16.
+//  Copyright © 2016 Josh Doman. All rights reserved.
+//
+
+import UIKit
+
+extension NSMutableAttributedString {
+    @discardableResult func bold(_ text: String, size: CGFloat) -> NSMutableAttributedString {
+        let attrs: [NSAttributedString.Key: Any] = [.font: UIFont.systemFont(ofSize: size, weight: .bold)]
+        let boldString = NSMutableAttributedString(string: text, attributes: attrs)
+        append(boldString)
+        return self
+    }
+    
+    @discardableResult func weighted(_ text: String, weight: UIFont.Weight, size: CGFloat) -> NSMutableAttributedString {
+        let attrs: [NSAttributedString.Key: Any] = [.font: UIFont.systemFont(ofSize: size, weight: weight)]
+        let boldString = NSMutableAttributedString(string: text, attributes: attrs)
+        append(boldString)
+        return self
+    }
+    
+    @discardableResult public func weightedColored(_ text: String, weight: UIFont.Weight, color: UIColor, size: CGFloat) -> NSMutableAttributedString {
+        let attrs: [NSAttributedString.Key: Any] = [.font: UIFont.systemFont(ofSize: size, weight: weight), NSAttributedString.Key.foregroundColor: color]
+        let boldString = NSMutableAttributedString(string: text, attributes: attrs)
+        append(boldString)
+        return self
+    }
+    
+    @discardableResult func colored(_ text: String, color: UIColor) -> NSMutableAttributedString {
+        let attrs: [NSAttributedString.Key: Any] = [NSAttributedString.Key.foregroundColor: color]
+        let colorString = NSMutableAttributedString(string: text, attributes: attrs)
+        append(colorString)
+        return self
+    }
+    
+    @discardableResult func normal(_ text: String) -> NSMutableAttributedString {
+        let normal = NSAttributedString(string: text)
+        append(normal)
+        return self
+    }
+}
