+++ PennMobileShared/Extensions/Padding.swift
@@ -1,0 +1,13 @@
+//
+//  Padding.swift
+//  PennMobileShared
+//
+//  Created by Josh Doman on 12/13/16.
+//  Copyright © 2016 Josh Doman. All rights reserved.
+//
+
+import CoreGraphics
+
+public class Padding {
+    public static let pad: CGFloat = 14.0
+}
