+++ PennMobileShared/Extensions/TimeZone+Extensions.swift
@@ -1,0 +1,13 @@
+//
+//  TimeZone+Extensions.swift
+//  PennMobileShared
+//
+//  Created by Josh Doman on 12/13/16.
+//  Copyright © 2016 Josh Doman. All rights reserved.
+//
+
+import Foundation
+
+public extension TimeZone {
+    static let nyc = TimeZone(identifier: "America/New_York")!
+}
