+++ PennMobileShared/Extensions/UIBarButtonItem+Extensions.swift
@@ -1,0 +1,22 @@
+//
+//  UIBarButtonItem+Extensions.swift
+//  PennMobileShared
+//
+//  Created by Josh Doman on 12/13/16.
+//  Copyright © 2016 Josh Doman. All rights reserved.
+//
+
+import UIKit
+
+extension UIBarButtonItem {
+    static func itemWith(colorfulImage: UIImage?, color: UIColor, target: AnyObject, action: Selector) -> UIBarButtonItem {
+        let button = UIButton(type: .custom)
+        button.setImage(colorfulImage?.withRenderingMode(UIImage.RenderingMode.alwaysTemplate), for: .normal)
+        button.tintColor = color
+        button.frame = CGRect(x: 0, y: 0, width: 30, height: 30)
+        button.addTarget(target, action: action, for: .touchUpInside)
+        
+        let barButtonItem = UIBarButtonItem(customView: button)
+        return barButtonItem
+    }
+}
