+++ PennMobileShared/Extensions/UIFont+Extensions.swift
@@ -1,0 +1,30 @@
+//
+//  UIFont+Extensions.swift
+//  PennMobileShared
+//
+//  Created by Josh Doman on 12/13/16.
+//  Copyright © 2016 Josh Doman. All rights reserved.
+//
+
+import UIKit
+
+extension UIFont {
+    public static let avenirMedium = UIFont.systemFont(ofSize: 21, weight: .regular)
+    public static let primaryTitleFont = UIFont.systemFont(ofSize: 21, weight: .semibold)
+    public static let secondaryTitleFont = UIFont.systemFont(ofSize: 11, weight: .medium)
+    
+    public static let interiorTitleFont = UIFont.systemFont(ofSize: 17, weight: .medium)
+    
+    public static let pollsTitleFont = UIFont.systemFont(ofSize: 17, weight: .semibold)
+    
+    public static let primaryInformationFont = UIFont.systemFont(ofSize: 14, weight: .semibold)
+    public static let secondaryInformationFont = UIFont.systemFont(ofSize: 14, weight: .regular)
+    
+    public static let footerDescriptionFont = UIFont.systemFont(ofSize: 10, weight: .regular)
+    public static let footerTransitionFont = UIFont.systemFont(ofSize: 10, weight: .semibold)
+    
+    public static let gsrTimeIncrementFont = UIFont.systemFont(ofSize: 20, weight: .semibold)
+    
+    public static let alertSettingsWarningFont = UIFont.systemFont(ofSize: 30, weight: .bold)
+    
+}
