+++ PennMobileShared/Extensions/URL+Extensions.swift
@@ -1,0 +1,33 @@
+//
+//  URL+Extensions.swift
+//  PennMobileShared
+//
+//  Created by Josh Doman on 12/13/16.
+//  Copyright © 2016 Josh Doman. All rights reserved.
+//
+
+import Foundation
+
+public extension URL {
+    mutating func appendQueryItem(name: String, value: String?) {
+        guard var urlComponents = URLComponents(string: absoluteString) else { return }
+        
+        var queryItems: [URLQueryItem] = urlComponents.queryItems ?? []
+        let queryItem = URLQueryItem(name: name, value: value)
+        
+        queryItems.append(queryItem)
+        
+        urlComponents.queryItems = queryItems
+        
+        self = urlComponents.url!
+    }
+    
+    var queryParameters: [String: String] {
+        guard
+            let components = URLComponents(url: self, resolvingAgainstBaseURL: true),
+            let queryItems = components.queryItems else { return [:] }
+        return queryItems.reduce(into: [String: String]()) { (result, item) in
+            result[item.name] = item.value
+        }
+    }
+}
